% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{beta.simu.plot}
\alias{beta.simu.plot}
\title{Histogram and Q-Q plot of simulated Beta distribution}
\usage{
beta.simu.plot(n, shape1, shape2, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{shape1}{non-negative parameters of the Beta distribution}

\item{shape2}{non-negative parameters of the Beta distribution}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Beta distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Beta distribution
}
\examples{
beta.simu.plot(n = 5, shape1 = 3, shape2 = 1, times = 100)
}

