% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{chisq.plot}
\alias{chisq.plot}
\title{Histogram and Q-Q plot of simulated Chi-Squared distribution}
\usage{
chisq.plot(n, df, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{df}{degrees of freedom (non-negative, but can be non-integer)}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Chi-Squared distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Chi-Squared distribution
}
\examples{
chisq.plot(n = 5, df = 4, times = 100)
}

