% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{coin}
\alias{coin}
\title{Theoretical Probability Distribution of Flipping Coins}
\usage{
coin(n, prob = NULL)
}
\arguments{
\item{n}{number of trials}

\item{prob}{probability assigned to each possible outcome}
}
\value{
Mean value and corresponding probabilities for all possible outcomes.
}
\description{
Mean and probability of flipping fair or loaded coin
}
\details{
The default probabilty equals to 1/n. All the assigned probabilites must between 0 and 1. The sum of
them equals to 1.
}
\examples{
coin(n = 4)
coin(6, c(0.1, 0.9))
}

