% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{coin.simu}
\alias{coin.simu}
\title{Probability Distribution of Simulated Coins Flipping}
\usage{
coin.simu(n, times, prob = NULL)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{times}{number of simulations}

\item{prob}{probability assigned to each possible outcome}
}
\value{
Mean value and corresponding probabilities for all simulated outcomes.
}
\description{
Mean and probability plot of flipping fair or loaded coin
}
\details{
The default probabilty equals to 1/n. All the assigned probabilites must between 0 and 1. The sum of
them equals to 1.
}
\examples{
coin.simu(n = 4, times = 1000)
coin.simu(4, 1000, prob = c(0.3, 0.7))
}

