% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{dice.plot}
\alias{dice.plot}
\title{Theoretical Probability Distribution Plot of Rolling Dice}
\usage{
dice.plot(n, prob = NULL, col = "black", type = NULL,
main = NULL, sub = NULL)
}
\arguments{
\item{n}{number of trials}

\item{prob}{probability assigned to each possible outcome}

\item{col}{color of the plot}

\item{type}{type of plot}

\item{main}{an overall title for the plot}

\item{sub}{a sub title for the plot}
}
\value{
Plot of mean value and corresponding probabilities for all possible outcomes.
}
\description{
Probability plot of rolling fair or loaded dice
}
\details{
The default probabilty equals to 1/n. All the assigned probabilites must between 0 and 1.
}
\examples{
dice.plot(n = 4, col ='red', type = 'p')
dice.plot(3, prob = c(0.3, 0.1, 0.2, 0.1, 0.1, 0.2))
}

