% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{dice}
\alias{dice}
\title{Theoretical Probability Distribution of Rolling Dice}
\usage{
dice(n, prob = NULL)
}
\arguments{
\item{n}{number of trials}

\item{prob}{probability assigned to each possible outcome}
}
\value{
Mean value and corresponding probabilities for all possible outcomes.
}
\description{
Mean and probability of rolling fair or loaded dice
}
\details{
The default probabilty equals to 1/n. All the assigned probabilites must between 0 and 1. The sum of
them equals to 1.
}
\examples{
dice(n = 4)
dice(2, c(0.1, 0.2, 0.2, 0.1, 0.3, 0.1))
}

