% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{expo.simu.plot}
\alias{expo.simu.plot}
\title{Histogram and Q-Q plot of simulated Exponential distribution}
\usage{
expo.simu.plot(n, rate = 1, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{rate}{vector of rates}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Exponential distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Exponential distribution
}
\examples{
expo.simu.plot(n = 5, rate = 2, times = 100)
}

