% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{expt.plot}
\alias{expt.plot}
\title{Theoretical Probability Distribution Plot of General Experiment}
\usage{
expt.plot(x, n, prob = NULL, col = "black", type = NULL,
main = NULL, sub = NULL)
}
\arguments{
\item{x}{vector, possible outcomes in one trial of experiment}

\item{n}{number of trials}

\item{prob}{probability assigned to each possible outcome}

\item{col}{color of the plot}

\item{type}{type of plot}

\item{main}{an overall title for the plot}

\item{sub}{a sub title for the plot}
}
\value{
Plot of mean value and corresponding probabilities for all possible outcomes.
}
\description{
General experiment plot with basic probability
}
\details{
The default probabilty equals to 1/n. All the assigned probabilites must between 0 and 1. The sum of
them equals to 1. The default probabilty equals to 1/n.
}
\examples{
expt.plot(x = c(1:3), n = 4, col ='red', type = 'p')
expt.plot(c(2:4), 3, prob = c(0.3, 0.5, 0.2))
}

