% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{gamm.simu.plot}
\alias{gamm.simu.plot}
\title{Histogram and Q-Q plot of simulated Gamma distribution}
\usage{
gamm.simu.plot(n, shape, rate = 1, scale = 1/rate, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{shape}{shape parameter}

\item{rate}{vector of rates}

\item{scale}{scale parameter}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Gamma distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Gamma distribution
}
\examples{
gamm.simu.plot(n = 5, shape = 3, rate = 1, times = 100)
}

