% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{geom.simu.plot}
\alias{geom.simu.plot}
\title{Histogram and Q-Q plot of simulated Geometric distribution}
\usage{
geom.simu.plot(n, prob, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{prob}{probability of success on each trial}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Geometric distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Geometric distribution
}
\examples{
geom.simu.plot(n = 10, prob = 0.2, times = 100)
}

