% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{hyper.plot}
\alias{hyper.plot}
\title{Histogram and Q-Q plot of simulated Hypergeometric distribution}
\usage{
hyper.plot(n, a, b, k, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{a}{the number of white balls in the urn}

\item{b}{the number of black balls in the urn}

\item{k}{the number of balls drawn from the urn}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Hypergeometric distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Hypergeometric distribution
}
\examples{
hyper.plot(n = 10, a = 10, b = 10, k = 5, times = 100)
}

