% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{nbinom.plot}
\alias{nbinom.plot}
\title{Histogram and Q-Q plot of simulated Negative Binomial distribution}
\usage{
nbinom.plot(n, size, prob, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of observations}

\item{size}{number of trials (zero or more)}

\item{prob}{probability of success on each trial}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Negative Binomial distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Negative Binomial distribution
}
\examples{
nbinom.plot(n = 10, size = 5, prob = 0.2, times = 100)
}

