% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/clttools.R
\name{pois.simu.plot}
\alias{pois.simu.plot}
\title{Histogram and Q-Q plot of simulated Poisson distribution}
\usage{
pois.simu.plot(n, lambda, times, ylim = NULL, qqplot = FALSE)
}
\arguments{
\item{n}{number of trials in one simulation}

\item{lambda}{parameter of Poisson distribution}

\item{times}{number of simulations}

\item{ylim}{range of y-axis}

\item{qqplot}{an argument to output Q-Q plot or not, can be TRUE or FALSE}
}
\value{
Histogram and Q-Q plot of simulated Poisson distribution, red curve represents theoretical density
}
\description{
Histogram and Q-Q plot of simulated Poisson distribution
}
\examples{
pois.simu.plot(n = 5, lambda = 3, times = 100)
}

